Removed interface TargetInstanceManager
