+++ webcurator-core/src/test/java/org/webcurator/core/scheduler/TargetInstanceManagerImplTest.java
@@ -12,14 +12,14 @@ import org.webcurator.core.util.TestAuditor;
 import org.webcurator.core.notification.*;
 import org.webcurator.domain.model.core.*;
 
-public class TargetInstanceManagerImplTest extends BaseWCTTest<TargetInstanceManagerImpl>{
+public class TargetInstanceManagerImplTest extends BaseWCTTest<TargetInstanceManager>{
 
 	private TargetInstanceDAO tiDao = null;
 	private ProfileDAO pDao = null;
 	
 	public TargetInstanceManagerImplTest()
 	{
-		super(TargetInstanceManagerImpl.class, 
+		super(TargetInstanceManager.class,
 				"/org/webcurator/core/scheduler/TargetInstanceManagerImplTest.xml");
 	}
